Guardar datos - reserva

+++ app/src/main/java/com/example/vuelav_app/Logico/Response/ReservaResponse.java
@@ -36,7 +36,7 @@ public class ReservaResponse  {
 
     private Long idTripulacion;
 
-    private Long idVuelo;
+    private int idVuelo;
 
     private Long idManifiesto;
